chore(ts-locals): use ts-locals on 123
