+++ lua/refactoring/ts-locals.lua
@@ -128,22 +128,22 @@ local function memoize(fn, hash_fn)
     return function(...)
         local key = hash_fn(...)
         if cache[key] == nil then
-            local v = { fn(...) } ---@type any
+            local v = { fn(...) } ---@type any[]
 
             for k, value in pairs(v) do
                 if value == nil then
-                    value[k] = vim.NIL
+                    value[k] = vim.NIL ---@type any
                 end
             end
 
             cache[key] = v
         end
 
-        local v = cache[key]
+        local v = cache[key] ---@type any[]
 
         for k, value in pairs(v) do
             if value == vim.NIL then
-                value[k] = nil
+                value[k] = nil --- @type any
             end
         end
 
@@ -190,16 +190,18 @@ end
 ---@return TSLocal[] definitions
 ---@return TSLocal[] references
 ---@return TSNode[] scopes
+---@return TSNode[] statements
 M.get = memoize(function(bufnr)
     local query, root = get_query(bufnr)
     if not query or not root then
-        return {}, {}, {}
+        return {}, {}, {}, {}
     end
 
     local definitions = {}
     local scopes = {}
     local references = {}
-    for id, node, metadata in query:iter_captures(root, bufnr) do
+    local statements = {}
+    for id, node, metadata in query:iter_captures(root, bufnr, 0, -1) do
         local kind = query.captures[id]
 
         local scope = "local" ---@type string
@@ -226,9 +228,13 @@ M.get = memoize(function(bufnr)
                 { kind = kind, node = node, scope = scope }
             )
         end
+
+        if node and kind == "statement" then
+            table.insert(statements, node)
+        end
     end
 
-    return definitions, references, scopes
+    return definitions, references, scopes, statements
 end, function(bufnr)
     local root = get_root(bufnr)
     if not root then
@@ -236,6 +242,13 @@ end, function(bufnr)
     end
     return tostring(root:id())
 end)
+
+---@param bufnr integer
+---@return TSNode[] statements
+M.get_statements = function(bufnr)
+    local _, _, _, statements = M.get(bufnr)
+    return statements
+end
 
 -- Get a single dimension table to look definition nodes.
 -- Keys are generated by using the range of the containing scope and the text of the definition node.
@@ -256,6 +269,7 @@ M.get_definitions_lookup_table = memoize(function(bufnr)
         return {}
     end
 
+    ---@type table<string, TSLocal>
     local result = {}
     for _, definition in ipairs(definitions) do
         for _, node_entry in ipairs(M.get_local_nodes(definition)) do
@@ -292,6 +306,7 @@ end)
 ---@param node TSNode: the definition node
 ---@param bufnr integer: the buffer
 ---@param scope_type TSScope: the scope type
+---@return TSNode[]
 function M.get_definition_scopes(node, bufnr, scope_type)
     local scopes = {}
     local scope_count = 1 ---@type integer|nil
@@ -361,12 +376,14 @@ function M.find_usages(node, scope_node, bufnr)
         return {}
     end
 
-    for id, node_capture in query:iter_captures(scope_node, bufnr) do
+    local definition = M.find_definition(node, bufnr)
+    for id, node_capture in query:iter_captures(scope_node, bufnr, 0, -1) do
         local kind = query.captures[id]
         if
             node_capture
             and kind == "reference"
             and ts.get_node_text(node_capture, bufnr) == node_text
+            and not node_capture:equal(definition)
         then
             table.insert(usages, node_capture)
         end
@@ -377,7 +394,7 @@ end
 
 ---@param node TSNode
 ---@param bufnr? integer
----@param allow_scope? boolean
+---@param allow_scope? boolean whether to return the same node if no scope is found
 ---@return TSNode|nil
 function M.containing_scope(node, bufnr, allow_scope)
     bufnr = bufnr or api.nvim_get_current_buf()
